downloadAzureStorageData/requirements.txt
azure-identity==1.6.1
azure-storage-blob==12.9.0b1